Fixed abs

+++ modules/task_3/nazarov_v_global_optimization/main.cpp
@@ -19,9 +19,9 @@ TEST(Global_Optimization_MPI, Test_First_Function) {
         std::cout << "Final result sequential, time to solve: " << cur-prev << std::endl;
         std::cout << "x: " << resEq.x << " y: " << resEq.y << " z: " << resEq.z << std::endl;
         bool equals = false;
-        if (std::abs(resEq.x - res.x) <= 0.001)
-            if (std::abs(resEq.y - res.y) <= 0.001)
-                if (std::abs(resEq.z - res.z) <= 0.001)
+        if (std::abs(static_cast<double>(resEq.x - res.x)) <= 0.001)
+            if (std::abs(static_cast<double>(resEq.y - res.y)) <= 0.001)
+                if (std::abs(static_cast<double>(resEq.z - res.z)) <= 0.001)
                     equals = true;
         ASSERT_TRUE(equals);
     }
